Added GENERATOR_EXTRA_FLAGS variable.
Building the CSS requires gecode headers and libraries. As the generator only parses correct c++ code then we have to include a path to the gecode headers (-I/../gecode) every time we call the generator. This variable is empty when the CSS is not build and has the appropriate inclusion options when it is.